work around automatic window renaming
When a root shell is started in tmux every command (even a single
return) changes the window title back to the current working directory.
This fact is very annoying because it means that properly named windows
will always revert their name under such circumstances.
This change introduces a patch that works around this issue.
Essentially, once the user has changed the window's name, all automatic
renaming is prohibited from taking place.

Note that it is still unclear as to why only windows where a new shell
is started indirectly (by means of the su command, for instance) seem to
have this issue, and not all windows. The reason seems to be some shell
interna but the exact details are unknown.