reduce memory footprint by employing malloc_trim
When tmux is used in conjunction with a lot of sessions/windows with a
potentially huge history, memory usage by tmux can become excessive.
Apparently, this it's less the fact that tmux is a memory hog or that it
actually leaks but rather that glibc's malloc fails to free the memory
back to the system in a timely manner.
This change employs malloc_trim(3) to advice glibc to free memory back
to the system in an attempt to reduce the overall memory footprint. The
patch was originally devised by 42wim [1].

[1] http://blog.42.be/2015/02/tmux-memory-usage-on-linux.html