Remove unused 'jump_again' parameter from window_copy_cursor_jump_to*
The functions of the window_copy_search* family accept a parameter,
'jump_again', influencing some of the internal logic.
As it turns out, all clients of these functions set this parameter to 1,
effectively rendering the additional logic for the conditional handling
of this flag superfluous. This change removes this parameter from all
functions and adjusts the callsites accordingly.