make jump-forward and jump-backward honor line wraps
The 'jump-forward', 'jump-backward', 'jump-to-forward', and
'jump-to-backward' primitives provided by tmux are not aware of line
wraps. That is, jumping to a certain character would stop at the end of
the line even if this line is actually wrapped around because the
terminal is too small.
This patch adjusts those commands to work properly in the context of
line wraps.