Remove unused 'moveflag' parameter from window_copy_search*
The functions of the window_copy_search* family accept a parameter,
'moveflag', that determines whether to search for a string at the
current cursor position or not.
As it turns out, all clients of these functions set this parameter to 1,
effectively rendering the additional logic for the conditional handling
of this flag superfluous. This change removes this parameter from all
functions and adjusts the callsites accordingly.