Fix repeated-search direction
It is utterly confusing and counter to how many other programs behave
to make repeated searches in the command output stateful in such a way
that search direction depends on the direction of the initial search.
With this change we change the behavior of the search-again and
search-reverse commands to always search in upward or downward
direction, respectively. As a result, the search-reverse command is now
a misnomer, but that is acceptable for a one-off patch, and certainly
less invasive than a fully blown rename or the introduction of a new
command.